// src/models/types.ts
export enum TipoAeronave {
    COMERCIAL = "COMERCIAL",
    MILITAR = "MILITAR"
}

export enum TipoPeca {
    NACIONAL = "NACIONAL",
    IMPORTADA = "IMPORTADA"
}

export enum TipoTeste {
    ELETRICO = "ELETRICO",
    HIDRAULICO = "HIDRAULICO",
    AERODINAMICO = "AERODINAMICO"
}

export enum Cargo {
    TECNICO = "TECNICO",
    ENGENHEIRO = "ENGENHEIRO",
    SUPERVISOR = "SUPERVISOR",
    GERENTE = "GERENTE"
}

export enum TipoEtapa {
    INSPECAO = "INSPECAO",
    MANUTENCAO = "MANUTENCAO",
    REPARO = "REPARO",
    TESTE = "TESTE"
}